+++ configs.ts
@@ -1,0 +1,11 @@
+import { VERSION } from './src/version';
+
+export const configs = {
+	'version': VERSION,
+	'plugin.config-persist.key': 'bluebase-config-test',
+
+	'foo': 5,
+
+	'plugin.config-persist.encryptionEnable': true,
+	'plugin.config-persist.encryptionKey': '123',
+};