+++ typings/@storybook/react/index.d.ts
@@ -1,2 +1,0 @@
-declare module '@storybook/react';
-declare var module: any;